Add login code in handleSubmit
if handle submit is clicked and auth is  in login, the signin service is
executed, when success save SignInOrupResponse in local storage.if fails
show an notification with message.

add loading to the form button.

add <Notification /> from mantine/core to show it when there some error
from the service.